+++ web/app/themes/welcome.helsinki/app/Blocks/Highlight.php
@@ -19,6 +19,7 @@ class Highlight extends NativeBlock
             'textColor' => $this->attributes->textColor,
             'linkText' => $this->attributes->linkText,
             'linkUrl' => $this->attributes->linkUrl,
+            'linkTargetBlank' => $this->attributes->linkTargetBlank,
         ];
     }
 }